requirements.txt
flask
Flask-SQLAlchemy
python-dotenv
mysqlclient
flask_migrate
